+++ common/src/main/kotlin/com/cobblemon/mod/common/CobblemonItems.kt
@@ -1059,8 +1059,8 @@ object CobblemonItems : PlatformRegistry<Registry<Item>, ResourceKey<Registry<It
     val CHOICE_SCARF = heldItem("choice_scarf")
     @JvmField
     val CLEANSE_TAG = heldItem("cleanse_tag")
-   // @JvmField
-   // val CLEAR_AMULET = heldItem("clear_amulet")
+    @JvmField
+    val CLEAR_AMULET = heldItem("clear_amulet")
     @JvmField
     val COVERT_CLOAK = heldItem("covert_cloak")
     @JvmField
@@ -1089,16 +1089,16 @@ object CobblemonItems : PlatformRegistry<Registry<Item>, ResourceKey<Registry<It
     val FOCUS_SASH = heldItem("focus_sash")
     @JvmField
     val GRASSY_SEED = heldItem("grassy_seed")
-   // @JvmField
-   // val GRIP_CLAW = heldItem("grip_claw")
+    @JvmField
+    val GRIP_CLAW = heldItem("grip_claw")
     @JvmField
     val HARD_STONE = heldItem("hard_stone")
     @JvmField
     val HEAVY_DUTY_BOOTS = heldItem("heavy_duty_boots")
     @JvmField
     val IRON_BALL = heldItem("iron_ball")
-   // @JvmField
-   // val LAGGING_TAIL = heldItem("lagging_tail")
+    @JvmField
+    val LAGGING_TAIL = heldItem("lagging_tail")
     @JvmField
     val LEFTOVERS = heldItem("leftovers")
     @JvmField
@@ -1111,8 +1111,8 @@ object CobblemonItems : PlatformRegistry<Registry<Item>, ResourceKey<Registry<It
     val LOADED_DICE = heldItem("loaded_dice")
     @JvmField
     val LUCKY_EGG = heldItem("lucky_egg")
-   // @JvmField
-   // val LUMINOUS_MOSS = heldItem("luminous_moss")
+    @JvmField
+    val LUMINOUS_MOSS = heldItem("luminous_moss")
     @JvmField
     val MAGNET = heldItem("magnet")
     @JvmField