+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fixed Moon Ball moon phase logic to actually work correctly
 - Added `translucent_cull` boolean option into resolver's layer to allow for translucent textures with culling
 - Added [LambDynamicLights](https://modrinth.com/mod/lambdynamiclights) support for items held by Pokémon.
+- Added the Clear Amulet, Grip Claw, Lagging Tail, and Luminous Moss
 
 ### Pokémon Added